personal-ai/personal-chatbot-voice/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-voice-chatbot"
version = "0.1.0"
description = "A personal AI voice assistant with voice cloning capabilities"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "openai>=1.0.0",
    "elevenlabs>=0.2.0",
    "whisper>=1.1.0",
    "pyttsx3>=2.90",
    "pydub>=0.25.1",
    "numpy>=1.24.0",
    "pyaudio>=0.2.11",
    "soundfile>=0.12.1",
    "sounddevice>=0.4.6",
    "webrtcvad>=2.0.10",
    "chromadb>=0.4.0",
    "sqlite-utils>=3.35.0",
    "langchain>=0.0.300",
    "sentence-transformers>=2.2.2",
    "scikit-learn>=1.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "colorlog>=6.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
    "pre-commit",
]
voice-cloning = [
    "TTS>=0.21.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
offline = [
    "vosk>=0.3.45",
    "edge-tts>=6.1.9",
]

[project.scripts]
voice-chatbot = "src.main:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88